Rename libs directory.
