+++ src/lib/AudioTool.js
