+++ src/lib/chartConfig.js
