+++ src/lib/getAverageVolume.js
