+++ src/lib/makeWeekScale.js
