+++ src/lib/sound.js
